trackproj/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.8.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
google-auth==2.43.0
google-genai==1.52.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
mysqlclient==2.2.7
numpy==2.3.5
packaging==25.0
pillow==11.2.1
portalocker==3.2.0
protobuf==6.33.1
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.9.0
pyOpenSSL==25.3.0
python-dotenv==1.1.0
qdrant-client==1.16.1
redis==7.1.0
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.3
tenacity==9.1.2
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
websockets==15.0.1
zope.interface==8.1.1
